Eliminate the FileUploadPage
